@@ -14,9 +14,11 @@ var RevisionInfo = require('./revisionInfo');
 var ValueTransformer = require('../../util/valueTransformer').ValueTransformer;
 var LayoutSwitcher = require('./layoutSwitcher');
 var ViewsFactory = require('./viewsFactory');
+var Loader = require('../common/loader.jsx');
+var LoaderMixin = require('../../mixins/LoaderMixin');
 
 var AssetView = React.createClass({
-    mixins:[Router.State, Flux.mixins.storeListener],
+    mixins:[Router.State, Flux.mixins.storeListener, LoaderMixin],
 
     getInitialState: function() {
         return {
@@ -25,16 +27,16 @@ var AssetView = React.createClass({
     },
 
     componentWillMount: function() {
-        var params = _.extend({}, this.getParams(), this.getQuery());
-        this.props.actions.loadAsset(params);
+        var params = _.extend({}, this.context.router.getCurrentParams(), this.context.router.getCurrentQuery());
+        this.waitFor(this.props.actions.loadAsset(params));
     },
 
     onAssetDelete: function() {
-        this.props.actions.deleteAsset(this.getParams());
+        this.props.actions.deleteAsset(this.context.router.getCurrentParams());
     },
 
     onAssetRestore: function() {
-        this.props.actions.restoreAsset(this.getParams());
+        this.props.actions.restoreAsset(this.context.router.getCurrentParams());
     },
 
     storeDidChange: function (storeName) {
@@ -64,7 +66,7 @@ var AssetView = React.createClass({
         var asset = assetStore.asset;
         var linkedAssets = assetStore.relatedAssets;
         var taxonomyPath = assetStore.taxonomyPath;
-        
+
         var assetLinks = linkedAssets.filter(function(e) { return e.assets != null }).map((entity) => {
             var links = entity.assets.map(function(asset){
                 return <Link className="nav-block__item-related"
@@ -88,71 +90,74 @@ var AssetView = React.createClass({
 
         var ViewComponent = ViewsFactory.getViewComponent(
             asset.screens, this.state.selectedScreen);
-        
+
         return (
             <div>
-                <SearchResultsHeader actions={this.props.actions} />
+                <SearchResultsHeader actions={this.props.actions} dispatcher={this.props.dispatcher} />
                 <h1 className="page-title">
                     <span className={titleClasses}>
                         {asset.name}
                     </span>
                 </h1>
                 <RevisionInfo asset={asset} dateTransform={dateTransform} />
-                <div className="grid">
-                    <div className="grid__item two-twelfths">
-                        <LayoutSwitcher 
-                            screens={asset.screens} 
-                            selectedScreen={this.state.selectedScreen}
-                            onChange={this.onScreenChange} />
-                        <TaxonomyPath taxonomyPath={taxonomyPath} />
-                        <nav className="nav-block">
-                            <span className="nav-block__title nav-block__title_type_second">Linked assets</span>
-                            <div className="nav-block__item">
-                                {assetLinks}
-                            </div>
-                        </nav>
-                        <nav className="nav-block">
-                            <span className="nav-block__title nav-block__title_type_second">Search result report</span>
-                            <ul className="nav-block__list">
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>Default Reports</span>
-                                </li>
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>Report with child assets</span>
-                                </li>
-                            </ul>
-                        </nav>
-                        <nav className="nav-block">
-                            <span className="nav-block__title nav-block__title_type_second">Export</span>
-                            <ul className="nav-block__list">
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>.txt</span>
-                                </li>
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>.xml</span>
-                                </li>
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>.doc</span>
-                                </li>
-                                <li className="nav-block__item">
-                                    <span className="link link_second"><span className="icon icon_download"></span>.zip all</span>
-                                </li>
-                            </ul>
-                        </nav>
+                <Loader loading={this.state.loading}>
+                    <div className="grid">
+                        <div className="grid__item two-twelfths">
+                            <LayoutSwitcher
+                                screens={asset.screens}
+                                selectedScreen={this.state.selectedScreen}
+                                onChange={this.onScreenChange} />
+                            <TaxonomyPath taxonomyPath={taxonomyPath} />
+                            <nav className="nav-block">
+                                <span className="nav-block__title nav-block__title_type_second">Linked assets</span>
+                                <div className="nav-block__item">
+                                    {assetLinks}
+                                </div>
+                            </nav>
+                            <nav className="nav-block">
+                                <span className="nav-block__title nav-block__title_type_second">Search result report</span>
+                                <ul className="nav-block__list">
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>Default Reports</span>
+                                    </li>
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>Report with child assets</span>
+                                    </li>
+                                </ul>
+                            </nav>
+                            <nav className="nav-block">
+                                <span className="nav-block__title nav-block__title_type_second">Export</span>
+                                <ul className="nav-block__list">
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>.txt</span>
+                                    </li>
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>.xml</span>
+                                    </li>
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>.doc</span>
+                                    </li>
+                                    <li className="nav-block__item">
+                                        <span className="link link_second"><span className="icon icon_download"></span>.zip all</span>
+                                    </li>
+                                </ul>
+                            </nav>
+                        </div>
+                        <div className="grid__item ten-twelfths">
+                            <ViewComponent
+                                screen={this.state.selectedScreen || {panels: []}}
+                                actions={this.props.actions}
+                                dispatcher={this.props.dispatcher}
+                                assetTypeId={asset.assetTypeId} />
+                            <AssetToolbar isHistory={asset.isHistory}
+                                          isDeleted={asset.isDeleted}
+                                          canEdit={asset.editable}
+                                          canDelete={asset.deletable}
+                                          onAssetDelete={this.onAssetDelete}
+                                          onAssetRestore={this.onAssetRestore} />
+                        </div>
                     </div>
-                    <div className="grid__item ten-twelfths">
-                        <ViewComponent
-                            screen={this.state.selectedScreen || {panels: []}}
-                            actions={this.props.actions}
-                            assetTypeId={asset.assetTypeId} />
-                        <AssetToolbar isHistory={asset.isHistory}
-                                      isDeleted={asset.isDeleted}
-                                      canEdit={asset.editable}
-                                      canDelete={asset.deletable}                                      
-                                      onAssetDelete={this.onAssetDelete}
-                                      onAssetRestore={this.onAssetRestore} />
-                    </div>
-                </div>
+                </Loader>
             </div>
         );
         }
