@@ -10,8 +10,8 @@ var AssetViewType2 = React.createClass({
 
     },
     render: function() {
-        var panels = this.props.screen.panels.map(function(el) {
-            return <Panel data={el} title={el.name}/>
+        var panels = this.props.screen.panels.map((el) => {
+            return <Panel data={el} title={el.name} dispatcher={this.props.dispatcher}/>
         });
 
         return (
