@@ -6,8 +6,8 @@ var TabPane = require('react-bootstrap/TabPane');
 
 var Panel = React.createClass({
     render: function() {
-        var attrs = this.props.data.attributes.map(function(attr) {
-            return AttributesFactory.getViewAttribute(attr.datatype, attr);
+        var attrs = this.props.data.attributes.map((attr) => {
+            return AttributesFactory.getViewAttribute(attr.datatype, attr, this.props.dispatcher);
         });
         return (
             <div>
